requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.48.1
kiwisolver==1.4.5
matplotlib==3.8.2
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyAudio==0.2.14
PyDejavu-Rollong==0.1.6
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.12.0
six==1.16.0
